+++ NEWS.md
@@ -13,6 +13,7 @@ TODO:
 
 # clusterProfiler 4.19.4.006
 
++ optimize `enrichGO()` to avoid memory boom when keyType is not ENTREZID (2026-01-21, Wed, #805)
 + `gson_GO_local()` to support local GO annotation by adding ancestral terms (2026-01-21, Wed)
 + `interpret()` implements a gene-based fallback mode for clusters with no enriched pathways, ensuring comprehensive analysis (2026-01-20, Tue)
 + `plot()` method for interpretation object to visualize the LLM-inferred regulatory network using `ggtangle` (2026-01-20, Tue)